requirements.txt
torch
transformers
sentencepiece
peft
datasets
faiss-cpu
sentence-transformers
